refactor(auth): enforce single auth and wallet hook import surface
- remove legacy duplicate AuthContext implementation

- switch app code to import useAuth/useWallet from '@/contexts' barrel

- enforce no-restricted-imports for direct AuthContext/WalletProvider paths

- document allowed hook import paths in CONTRIBUTING.md

Closes #146
