+++ src/app/login/layout.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from "react";
-import { AuthProvider } from "@/context/AuthContext";
+import { AuthProvider } from "@/contexts";
 
 export default function LoginLayout({ children }: { children: React.ReactNode }) {
   return (
