+++ src/components/WalletConnectButton.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useState, useEffect } from 'react';
-import { useWallet } from '../contexts';
+import { useWallet } from '@/contexts';
 
 const WalletIcon = () => (
   <svg className="w-5 h-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
