+++ src/components/auth/ProtectedRoute.tsx
@@ -14,7 +14,7 @@
 
 import { useEffect } from "react";
 import { useRouter } from "next/navigation";
-import { useAuth } from "@/contexts/AuthContext";
+import { useAuth } from "@/contexts";
 import { SIGN_IN_PATH } from "@/lib/auth-constants";
 
 interface ProtectedRouteProps {
